Make table layout fixed
Each column has variable width and that doesn't work well with
Sortable. When dragging and dropping the issue can appear far away from
the pointer and be hard to drop at the correct spot. This problem goes
away by using a fixed tabled layout and also looks nicer since the
table does not redraw just because an issue is moved.