Extract just category ID and name in the controller
Some entities have a different ID number from what we want when we make
the bulk_update call. In particular assignee should be the user_id
field, not id. Instead of just making a list of the categories, make a
new list where we only add the ID and name and nothing else so that
it's easy and straight forward to use them and allows us to remove some
of the checks we did in the view.